Corrections related to wrong assertions and update of readme (#17)
* Update of the config now optional+ possibility to retrieve baked wfs

Baking now allows a simple retrieval of baked waveforms, that can be created according under the constraint of being of a certain length caught from another baked waveform (to ensure compatibility with overriding waveforms with add_compiled feature). Using this functionality is done by inserting in the new baking object a "baking_index", which will identify the constraining length for dealing with overrides syntax for add_compiled. Note that introducing this baking index induces that the waveform baked is no longer updated in the config, it is accessible directly in a overrides format using the method b.get_waveforms_dict.

* Added example for pre_compile usage + changed the input parameters for b

Instead of declaring a matching length constraint, one has now to indicate a baking index that indicates which baked waveform should be overwritten (and ensure it matches its length for overriding it with add_compiled)

* Reformat

* Added method delete_baked_Op

This new method (called using b.delete_baked_Op(qe)) allows the direct removal from the input config of the baking object the associated baked operation and associated waveforms

* Modified input tests as assertion tests and corrected wrong behavior

Corrected issue for amplitude modulation + updated readme + added test

Corrected assertions

* Removed update_config parameter and corrected test

* Added more tests, and extended usage of align and delete_baked_op